Migrate to tsdown.
# Conflicts:
#	package-lock.json
#	package.json

# Conflicts:
#	package-lock.json
#	package.json
