+++ packed/esm/package.json
@@ -6,9 +6,9 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "@flystorage/file-storage": "file:~/Sites/filestorage/packed/flystorage-file-storage-0.1.7.tgz",
-    "@flystorage/local-fs": "file:~/Sites/filestorage/packed/flystorage-local-fs-0.1.10.tgz",
-    "@flystorage/stream-mime-type": "file:~/Sites/filestorage/packed/flystorage-stream-mime-type-0.1.7.tgz"
+    "@flystorage/file-storage": "file:~/Sites/filestorage/packed/flystorage-file-storage-1.2.0.tgz",
+    "@flystorage/local-fs": "file:~/Sites/filestorage/packed/flystorage-local-fs-1.2.0.tgz",
+    "@flystorage/stream-mime-type": "file:~/Sites/filestorage/packed/flystorage-stream-mime-type-1.1.0.tgz"
   },
   "keywords": [],
   "author": "",
